Extract inner class into independant class
